abstracted uploaded file and customer uploaded file
