+++ packages/framework/src/Component/AbstractUploadedFile/AbstractUploadedFile.php
@@ -1,0 +1,220 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shopsys\FrameworkBundle\Component\AbstractUploadedFile;
+
+use DateTime;
+use Doctrine\ORM\Mapping as ORM;
+use Shopsys\FrameworkBundle\Component\FileUpload\EntityFileUploadInterface;
+use Shopsys\FrameworkBundle\Component\FileUpload\Exception\InvalidFileKeyException;
+use Shopsys\FrameworkBundle\Component\FileUpload\FileForUpload;
+use Shopsys\FrameworkBundle\Component\FileUpload\FileNamingConvention;
+use Shopsys\FrameworkBundle\Component\String\TransformString;
+
+/**
+ * @ORM\MappedSuperclass
+ */
+abstract class AbstractUploadedFile implements EntityFileUploadInterface, UploadedFileInterface
+{
+    /**
+     * @var string|null
+     */
+    protected $temporaryFilename;
+
+    /**
+     * @var string
+     * @ORM\Column(type="string", length=5)
+     */
+    protected $extension;
+
+    /**
+     * @var \DateTime
+     * @ORM\Column(type="datetime")
+     */
+    protected $modifiedAt;
+
+    /**
+     * @var string
+     * @ORM\Column(type="string", length=255)
+     */
+    protected $name;
+
+    /**
+     * @var int
+     * @ORM\Column(type="integer")
+     * @ORM\Id
+     * @ORM\GeneratedValue(strategy="IDENTITY")
+     */
+    protected $id;
+
+    /**
+     * @var string
+     * @ORM\Column(type="string", length=255)
+     */
+    protected $slug;
+
+    /**
+     * @return string
+     */
+    public function getSlugWithExtension(): string
+    {
+        return $this->slug . '.' . $this->extension;
+    }
+
+    /**
+     * @return string
+     */
+    public function getTemporaryFilename(): string
+    {
+        return $this->temporaryFilename;
+    }
+
+    /**
+     * @return string
+     */
+    public function getFilename(): string
+    {
+        return $this->id . '.' . $this->extension;
+    }
+
+    /**
+     * @param string $key
+     * @param string $originalFilename
+     */
+    public function setFileAsUploaded(string $key, string $originalFilename): void
+    {
+        if ($key !== $this::getUploadKey()) {
+            throw new InvalidFileKeyException($key);
+        }
+
+        $this->extension = pathinfo($originalFilename, PATHINFO_EXTENSION);
+    }
+
+    /**
+     * @param string $key
+     */
+    public function setFileKeyAsUploaded(string $key): void
+    {
+        if ($key !== $this::getUploadKey()) {
+            throw new InvalidFileKeyException($key);
+        }
+
+        $this->temporaryFilename = '';
+    }
+
+    /**
+     * @param string $temporaryFilename
+     */
+    public function setNameAndSlug(string $temporaryFilename): void
+    {
+        $filenameWithoutExtension = pathinfo($temporaryFilename, PATHINFO_FILENAME);
+
+        $this->setName($filenameWithoutExtension);
+        $this->setSlug(TransformString::stringToFriendlyUrlSlug($filenameWithoutExtension));
+    }
+
+    /**
+     * @param string $name
+     */
+    public function setName($name): void
+    {
+        $this->name = $name;
+    }
+
+    /**
+     * @return string
+     */
+    public function getNameWithExtension(): string
+    {
+        return $this->name . '.' . $this->extension;
+    }
+
+    /**
+     * @return string
+     */
+    public function getExtension()
+    {
+        return $this->extension;
+    }
+
+    /**
+     * @return int|null
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    /**
+     * @param string $slug
+     */
+    public function setSlug($slug): void
+    {
+        $this->slug = $slug;
+    }
+
+    /**
+     * @return string
+     */
+    public function getSlug()
+    {
+        return $this->slug;
+    }
+
+    /**
+     * @return string
+     */
+    public function getName()
+    {
+        return $this->name;
+    }
+
+    /**
+     * @return \DateTime
+     */
+    public function getModifiedAt()
+    {
+        return $this->modifiedAt;
+    }
+
+    /**
+     * @return string
+     */
+    abstract protected function getUploadKey(): string;
+
+    /**
+     * @return string
+     */
+    abstract protected function getFileForUploadCategory(): string;
+
+    /**
+     * @return \Shopsys\FrameworkBundle\Component\FileUpload\FileForUpload[]
+     */
+    public function getTemporaryFilesForUpload(): array
+    {
+        $files = [];
+
+        if ($this->temporaryFilename !== null) {
+            $files[$this->getUploadKey()] = new FileForUpload(
+                $this->temporaryFilename,
+                static::class,
+                $this->getFileForUploadCategory(),
+                null,
+                FileNamingConvention::TYPE_ID,
+            );
+        }
+
+        return $files;
+    }
+
+    /**
+     * @param string $temporaryFilename
+     */
+    public function setTemporaryFilename(string $temporaryFilename): void
+    {
+        $this->temporaryFilename = $temporaryFilename;
+        // workaround: Entity must be changed so that preUpdate and postUpdate are called
+        $this->modifiedAt = new DateTime();
+    }
+}
