--- packages/framework/src/Component/CustomerUploadedFile/Config/Exception/CustomerUploadedFileConfigException.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Shopsys\FrameworkBundle\Component\CustomerUploadedFile\Config\Exception;
-
-use Shopsys\FrameworkBundle\Component\CustomerUploadedFile\Exception\CustomerFileException;
-
-interface CustomerUploadedFileConfigException extends CustomerFileException
-{
-}
