--- packages/framework/src/Component/CustomerUploadedFile/Config/Exception/CustomerUploadedFileConfigurationParseException.php
@@ -1,34 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Shopsys\FrameworkBundle\Component\CustomerUploadedFile\Config\Exception;
-
-use Exception;
-use Throwable;
-
-class CustomerUploadedFileConfigurationParseException extends Exception implements CustomerUploadedFileConfigException
-{
-    protected string $entityClass;
-
-    /**
-     * @param string $entityClass
-     * @param \Throwable|null $previous
-     */
-    public function __construct($entityClass, ?Throwable $previous = null)
-    {
-        $this->entityClass = $entityClass;
-
-        $message = sprintf('Parsing of config entity class "%s" failed.', $this->entityClass);
-
-        parent::__construct($message, 0, $previous);
-    }
-
-    /**
-     * @return string
-     */
-    public function getEntityClass()
-    {
-        return $this->entityClass;
-    }
-}
