+++ packages/framework/src/Component/CustomerUploadedFile/CustomerUploadedFile.php
@@ -4,14 +4,9 @@ declare(strict_types=1);
 
 namespace Shopsys\FrameworkBundle\Component\CustomerUploadedFile;
 
-use DateTime;
 use Doctrine\ORM\Mapping as ORM;
+use Shopsys\FrameworkBundle\Component\AbstractUploadedFile\AbstractUploadedFile;
 use Shopsys\FrameworkBundle\Component\CustomerUploadedFile\Exception\CustomerFileNotFoundException;
-use Shopsys\FrameworkBundle\Component\FileUpload\EntityFileUploadInterface;
-use Shopsys\FrameworkBundle\Component\FileUpload\Exception\InvalidFileKeyException;
-use Shopsys\FrameworkBundle\Component\FileUpload\FileForUpload;
-use Shopsys\FrameworkBundle\Component\FileUpload\FileNamingConvention;
-use Shopsys\FrameworkBundle\Component\String\TransformString;
 use Shopsys\FrameworkBundle\Model\Customer\User\CustomerUser;
 
 /**
@@ -22,29 +17,9 @@ use Shopsys\FrameworkBundle\Model\Customer\User\CustomerUser;
  * })
  * @ORM\Entity
  */
-class CustomerUploadedFile implements EntityFileUploadInterface
+class CustomerUploadedFile extends AbstractUploadedFile
 {
     protected const string UPLOAD_KEY = 'customerUploadedFile';
-
-    /**
-     * @var int
-     * @ORM\Column(type="integer")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="IDENTITY")
-     */
-    protected $id;
-
-    /**
-     * @var string
-     * @ORM\Column(type="string", length=255)
-     */
-    protected $name;
-
-    /**
-     * @var string
-     * @ORM\Column(type="string", length=255)
-     */
-    protected $slug;
 
     /**
      * @var string
@@ -60,26 +35,9 @@ class CustomerUploadedFile implements EntityFileUploadInterface
 
     /**
      * @var string
-     * @ORM\Column(type="string", length=5)
-     */
-    protected $extension;
-
-    /**
-     * @var \DateTime
-     * @ORM\Column(type="datetime")
-     */
-    protected $modifiedAt;
-
-    /**
-     * @var string
      * @ORM\Column(type="string", length=100)
      */
     protected $type;
-
-    /**
-     * @var string|null
-     */
-    protected $temporaryFilename;
 
     /**
      * @var int
@@ -128,133 +86,6 @@ class CustomerUploadedFile implements EntityFileUploadInterface
         $this->position = $position;
         $this->customerUser = $customerUser;
         $this->hash = $hash;
-    }
-
-    /**
-     * @return \Shopsys\FrameworkBundle\Component\FileUpload\FileForUpload[]
-     */
-    public function getTemporaryFilesForUpload(): array
-    {
-        $files = [];
-
-        if ($this->temporaryFilename !== null) {
-            $files[static::UPLOAD_KEY] = new FileForUpload(
-                $this->temporaryFilename,
-                self::class,
-                $this->entityName,
-                null,
-                FileNamingConvention::TYPE_ID,
-            );
-        }
-
-        return $files;
-    }
-
-    /**
-     * @param string $key
-     * @param string $originalFilename
-     */
-    public function setFileAsUploaded(string $key, string $originalFilename): void
-    {
-        if ($key !== static::UPLOAD_KEY) {
-            throw new InvalidFileKeyException($key);
-        }
-
-        $this->extension = pathinfo($originalFilename, PATHINFO_EXTENSION);
-    }
-
-    /**
-     * @param string $key
-     */
-    public function setFileKeyAsUploaded(string $key): void
-    {
-        if ($key !== static::UPLOAD_KEY) {
-            throw new InvalidFileKeyException($key);
-        }
-
-        $this->temporaryFilename = '';
-    }
-
-    /**
-     * @param string $temporaryFilename
-     */
-    public function setTemporaryFilename(string $temporaryFilename): void
-    {
-        $this->temporaryFilename = $temporaryFilename;
-        // workaround: Entity must be changed so that preUpdate and postUpdate are called
-        $this->modifiedAt = new DateTime();
-    }
-
-    /**
-     * @return string
-     */
-    public function getFilename(): string
-    {
-        return $this->id . '.' . $this->extension;
-    }
-
-    /**
-     * @return int
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    /**
-     * @return string
-     */
-    public function getName()
-    {
-        return $this->name;
-    }
-
-    /**
-     * @return string
-     */
-    public function getNameWithExtension(): string
-    {
-        return $this->name . '.' . $this->extension;
-    }
-
-    /**
-     * @param string $name
-     */
-    public function setName($name): void
-    {
-        $this->name = $name;
-    }
-
-    /**
-     * @return string
-     */
-    public function getTemporaryFilename(): string
-    {
-        return $this->temporaryFilename;
-    }
-
-    /**
-     * @return string
-     */
-    public function getSlug()
-    {
-        return $this->slug;
-    }
-
-    /**
-     * @return string
-     */
-    public function getSlugWithExtension(): string
-    {
-        return $this->slug . '.' . $this->extension;
-    }
-
-    /**
-     * @param string $slug
-     */
-    public function setSlug($slug): void
-    {
-        $this->slug = $slug;
     }
 
     /**
@@ -316,29 +147,26 @@ class CustomerUploadedFile implements EntityFileUploadInterface
     }
 
     /**
-     * @param string $temporaryFilename
-     */
-    public function setNameAndSlug(string $temporaryFilename): void
-    {
-        $filenameWithoutExtension = pathinfo($temporaryFilename, PATHINFO_FILENAME);
-
-        $this->setName($filenameWithoutExtension);
-        $this->setSlug(TransformString::stringToFriendlyUrlSlug($filenameWithoutExtension));
-    }
-
-    /**
-     * @return \DateTime
-     */
-    public function getModifiedAt()
-    {
-        return $this->modifiedAt;
-    }
-
-    /**
      * @return int
      */
     public function getPosition()
     {
         return $this->position;
+    }
+
+    /**
+     * @return string
+     */
+    protected function getUploadKey(): string
+    {
+        return self::UPLOAD_KEY;
+    }
+
+    /**
+     * @return string
+     */
+    protected function getFileForUploadCategory(): string
+    {
+        return $this->entityName;
     }
 }
