+++ packages/framework/src/Component/CustomerUploadedFile/Exception/CustomerFileException.php
@@ -4,8 +4,16 @@ declare(strict_types=1);
 
 namespace Shopsys\FrameworkBundle\Component\CustomerUploadedFile\Exception;
 
-use Throwable;
+use Exception;
 
-interface CustomerFileException extends Throwable
+class CustomerFileException extends Exception
 {
+    /**
+     * @param string $message
+     * @param \Exception|null $previous
+     */
+    public function __construct($message = '', ?Exception $previous = null)
+    {
+        parent::__construct($message, 0, $previous);
+    }
 }
