--- packages/framework/src/Component/CustomerUploadedFile/Exception/EntityIdentifierException.php
@@ -1,19 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Shopsys\FrameworkBundle\Component\CustomerUploadedFile\Exception;
-
-use Exception;
-
-class EntityIdentifierException extends Exception implements CustomerFileException
-{
-    /**
-     * @param string $message
-     * @param \Exception|null $previous
-     */
-    public function __construct($message = '', ?Exception $previous = null)
-    {
-        parent::__construct($message, 0, $previous);
-    }
-}
