+++ packages/framework/src/Component/UploadedFile/Config/UploadedFileConfig.php
@@ -6,7 +6,7 @@ namespace Shopsys\FrameworkBundle\Component\UploadedFile\Config;
 
 use Shopsys\FrameworkBundle\Component\UploadedFile\Config\Exception\UploadedFileEntityConfigNotFoundException;
 
-class UploadedFileConfig
+class UploadedFileConfig implements UploadedFileConfigInterface
 {
     /**
      * @param \Shopsys\FrameworkBundle\Component\UploadedFile\Config\UploadedFileEntityConfig[] $uploadedFileEntityConfigsByClass
