+++ packages/framework/src/Component/UploadedFile/UploadedFileLocator.php
@@ -4,43 +4,13 @@ declare(strict_types=1);
 
 namespace Shopsys\FrameworkBundle\Component\UploadedFile;
 
-use League\Flysystem\FilesystemOperator;
+use Shopsys\FrameworkBundle\Component\AbstractUploadedFile\AbstractUploadedFileLocator;
+use Shopsys\FrameworkBundle\Component\AbstractUploadedFile\UploadedFileInterface;
 use Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig;
-use Shopsys\FrameworkBundle\Component\Router\DomainRouterFactory;
 use Shopsys\FrameworkBundle\Component\UploadedFile\Exception\FileNotFoundException;
 
-class UploadedFileLocator
+class UploadedFileLocator extends AbstractUploadedFileLocator
 {
-    /**
-     * @param string $uploadedFileDir
-     * @param \League\Flysystem\FilesystemOperator $filesystem
-     * @param \Shopsys\FrameworkBundle\Component\Router\DomainRouterFactory $domainRouterFactory
-     */
-    public function __construct(
-        protected readonly string $uploadedFileDir,
-        protected readonly FilesystemOperator $filesystem,
-        protected readonly DomainRouterFactory $domainRouterFactory,
-    ) {
-    }
-
-    /**
-     * @param \Shopsys\FrameworkBundle\Component\UploadedFile\UploadedFile $uploadedFile
-     * @return string
-     */
-    public function getRelativeUploadedFileFilepath(UploadedFile $uploadedFile): string
-    {
-        return $this->getRelativeFilePath($uploadedFile->getFilename());
-    }
-
-    /**
-     * @param \Shopsys\FrameworkBundle\Component\UploadedFile\UploadedFile $uploadedFile
-     * @return string
-     */
-    public function getAbsoluteUploadedFileFilepath(UploadedFile $uploadedFile): string
-    {
-        return $this->getAbsoluteFilePath($uploadedFile->getFilename());
-    }
-
     /**
      * @param \Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig $domainConfig
      * @param \Shopsys\FrameworkBundle\Component\UploadedFile\UploadedFile $uploadedFile
@@ -61,31 +31,11 @@ class UploadedFileLocator
     }
 
     /**
-     * @param \Shopsys\FrameworkBundle\Component\UploadedFile\UploadedFile $uploadedFile
-     * @return bool
-     */
-    public function fileExists(UploadedFile $uploadedFile): bool
-    {
-        $fileFilepath = $this->getAbsoluteUploadedFileFilepath($uploadedFile);
-
-        return $this->filesystem->has($fileFilepath);
-    }
-
-    /**
-     * @param string $entityName
+     * @param \Shopsys\FrameworkBundle\Component\AbstractUploadedFile\UploadedFileInterface $uploadedFile
      * @return string
      */
-    protected function getRelativeFilePath(string $entityName): string
+    protected function getFilePath(UploadedFileInterface $uploadedFile): string
     {
-        return $entityName;
-    }
-
-    /**
-     * @param string $entityName
-     * @return string
-     */
-    public function getAbsoluteFilePath(string $entityName): string
-    {
-        return $this->uploadedFileDir . $this->getRelativeFilePath($entityName);
+        return $uploadedFile->getFilename();
     }
 }
